Fix Display Bug
Original liamcaReference used a span, which is "illegal" when a ul is
nested inside of it. Thus it would give a build warning. To fix that the
dropdown was switched to a div, but last update didn't include display:
inline-block in order to force element to behave like a span.