Update EntityBase
Changes the default type property in the entity base class to "docType"
so that an additional property "type" may be used in each class that
inherits from the base class. The docType property is set based on the
string name passed to the base class. Thus the docType in Trails will
still be set to "trail". Also adds the JsonProperty to all variables in
the EntityBase and Trail classes.