Update WebAPI to Start Adding Hash Partitioning
Fixes the _typePredicate to now resolve to the docType variable so that
a separate subtype can be applied to an entity. Also adds partially
completed code as part of demonstrating how to add Hash Partitioning for
multi tenants applications. The hashing sample is based on code from
Ryan CrawCour's DocumentDB MVA sample. Credits/license inclusions will
be updated accordingly.