Build 17
Add in trail index and management page to demonstrate interconnect with
DocumentDB via Web API. Will be expanded further with an implementation
for grabbing geojson objects during edit or creation of new trails.
Somewhat hackish changes in the WebAPI project to reflect the lack of
overt docType or Type reference in the sample trail data.