Add Partition Initializers
Adds sample initializers for the various partition resolvers with the
ability to pass in the DocumentClient as a parameter. Also adds an async
task for LogDocumentCountsPerCollection to the DocumentClientHelpers
class which will return a dictionary<string, int> named log with
collection.Id and numDocuments as key / value pairs. The initializers
have the OfferType set in the Web.config file.