Add VSXtoMD
Adds the following gist
(https://gist.github.com/formix/515d3d11ee7c1c252f92) to project files
for converting the auto-generated XML to markdown for github - you need
to configure it as needed. There is an error generated by the Vsxmd
nuget package (v1.1) when used with the WebAPI project. The error with
vsxmd is related to the "[assembly:" tags with Ninject. The section of
the VSXtoMD program that applies to that is commented out to allow
output.