Update NinjectWebCommon.cs
Moves the selection of a partition resolver for the sample item
repository to Ninject. Also shows some sample code on grabbing a value
from httpContext.Sessions. You could modify the NinjectWebCommon class
to grab the setting from the config file; however, that will be the
default option in future commits.