Build 21
Supports SignalR integration in WebAPI and WebRole projects to allow
real-time updates of the Trails data across multiple clients. Every
create, update, or delete operation against the DocumentDB endpoint
triggers a server side broadcast to all clients. You must have
websockets enabled on your server in order to have it work properly. If
you get a 503 error from Azure, it is because they are limiting the
number of instances that are allowed. A free web app is limited to 5
instances, whereas a Standard service place is capable of up to 35.